The TOC is not displayed for new style TR
The full TOC is copied onto the nav file, i.e., it is pretty much unnecessary. This may create problems with old EPUB readers that do not use the EPUB3 nav file; we will see if this is a problem in which case this step can be rolled back. It is a single line in the adjunct to the book.css file.